+++ src/Segment/ParsedRegExp.php
@@ -14,7 +14,7 @@ final class ParsedRegExp extends ParsedSegment
     public function __construct(
         string $value,
         string $raw,
-        ?ThrowMode $mode = null,
+        ?ThrowMode $mode,
         public bool $filterKeys,
     ) {
         $this->value = $this->raw = $value;
@@ -25,14 +25,24 @@ final class ParsedRegExp extends ParsedSegment
         // Note: PHP does not have a built-in way to validate regex syntax without executing it.
         // The following is a workaround to check if the regex is valid.
         // If the regex is invalid, preg_match() will either return false or throw an error.
+        set_error_handler(
+            callback: static function ($_, $msg): never { throw new \RuntimeException($msg); },
+            error_levels: E_WARNING
+        );
         try {
             // This will throw a SyntaxError if the regex is invalid
+            /** @psalm-suppress UnusedFunctionCall */
             if (false === preg_match($value, '')) {
+                // If preg_match returns false, it means there was an error in the regex
                 throw new SyntaxError('Invalid regular expression: '.$value.' - '.preg_last_error_msg());
             }
         } catch (\Throwable $e) {
+            $e instanceof SyntaxError && throw $e;
+
             // If the regular expression is invalid, we throw an error
             throw new SyntaxError('Invalid regular expression: '.$value.' - '.$e->getMessage(), 0, $e);
+        } finally {
+            restore_error_handler();
         }
     }
 }